+++ src/server/index.ts
@@ -24,6 +24,7 @@ import axios from 'axios'
 
 import persisted from '../../@generated/graphql/persisted.json'
 import storeConfig, { api } from '../../store.config'
+import secrets from '../../secrets.hidden.json'
 
 interface ExecuteOptions {
   operationName: string
@@ -40,6 +41,28 @@ type ShippingVariable = {
   postalCode: string
 }
 
+type WishlistVariable = {
+  email: string
+  productIds: string
+  id?: string
+}
+
+type GetWishListT = {
+  email: string
+}
+
+type GetWishListProductsT = {
+  productIds: string
+}
+interface InstallmentProps {
+  Value: number
+  InterestRate: number
+  TotalValuePlusInterestRate: number
+  NumberOfInstallments: number
+  Name: string
+  PaymentSystemName: string
+}
+
 // const typesArray = loadFilesSync('./src/server', {
 //   extensions: ['gql'],
 // })
@@ -47,7 +70,6 @@ type ShippingVariable = {
 // const typeDefsFromfile = mergeTypeDefs(typesArray)
 
 const persistedQueries = new Map(Object.entries(persisted))
-
 const apiOptions: APIOptions = {
   platform: storeConfig.platform as APIOptions['platform'],
   account: storeConfig.api.storeId,
@@ -209,12 +231,34 @@ const typeDefs = `
     messages: [MessageInfo]
   }
 
+  type WishlistData {
+    email: String
+    productIds: String
+    id: String
+  }
+
+  type WishlistDataMutation {
+    message: String
+  }
+
   input ShippingItem {
     id: String
     quantity: String
     seller: String
   }
-  
+
+  type WishListProductsData {
+    productId: String
+    productName: String
+    link: String
+    image: String
+    price: String
+    listPrice: String,
+    brand: String,
+    allInstallment: [Installment!]
+    discountHighlights: [DiscountHighlight]
+  }
+
   type Query {
     shipping(
       postalCode: String
@@ -224,12 +268,18 @@ const typeDefs = `
     ): ShippingData
 
     newsLetter(email:String!):NewsLetterData
+    getWishlist(email: String): WishlistData
+    getWishListProducts(productIds: String): [WishListProductsData]
   }
 
   type Mutation {
     newsLetterUpdate(email:String!, id:String): NewsLetterData
+    setWishlist(
+      email: String
+      productIds: String
+      id: String
+    ): WishlistDataMutation
   }
-
 `
 
 const resolvers = {
@@ -244,9 +294,12 @@ const resolvers = {
   },
   Query: {
     shipping,
+    getWishlist,
+    getWishListProducts,
     newsLetter,
   },
   Mutation: {
+    setWishlist,
     newsLetterUpdate,
   },
 }
@@ -342,6 +395,132 @@ async function shipping(
   return data
 }
 
+async function getWishlist(_: unknown, { email }: GetWishListT) {
+  const { data } = await axios.get(
+    `https://${api.storeId}.${api.environment}.com.br/api/dataentities/WS/search?email=${email}&_fields=_all`,
+    {
+      headers: {
+        'X-VTEX-API-AppKey': process.env.API_KEY ?? secrets.API_KEY,
+        'X-VTEX-API-AppToken': process.env.API_TOKEN ?? secrets.API_TOKEN,
+        'Content-Type': 'application/json',
+        Accept: 'application/json',
+        'REST-Range': 'resources=0-10',
+      },
+    }
+  )
+
+  if (!data) {
+    return new GraphQLError('Não foi possivel ver a wishlist desse usuario')
+  }
+
+  return data[0]
+}
+
+async function setWishlist(
+  _: unknown,
+  { email, productIds, id }: WishlistVariable
+) {
+  const { status } = await axios.put(
+    id
+      ? `https://${api.storeId}.${api.environment}.com.br/api/dataentities/WS/documents/${id}`
+      : `https://${api.storeId}.${api.environment}.com.br/api/dataentities/WS/documents`,
+    {
+      email,
+      productIds,
+    },
+    {
+      headers: {
+        'X-VTEX-API-AppKey': process.env.API_KEY ?? secrets.API_KEY,
+        'X-VTEX-API-AppToken': process.env.API_TOKEN ?? secrets.API_TOKEN,
+        'Content-Type': 'application/json',
+        Accept: 'application/json',
+      },
+    }
+  )
+
+  if (status === 201 || status === 200 || status === 204) {
+    return { message: 'Sucesso' }
+  }
+
+  return new GraphQLError('Não foi adicionar esse produto na Wishlist')
+}
+
+async function getWishListProducts(
+  _: unknown,
+  { productIds }: GetWishListProductsT
+) {
+  let path = ''
+
+  const ObjProductsIds = JSON.parse(productIds)
+
+  if (!ObjProductsIds) {
+    return new GraphQLError('Houve um erro com sua lista de produtos')
+  }
+
+  ObjProductsIds.forEach((refId: string) => {
+    path = `${path}&fq=productId:${refId}`
+  })
+
+  const { data } = await axios.get(
+    `https://${api.storeId}.${api.environment}.com.br/api/catalog_system/pub/products/search?_from=0&_to=19${path}`,
+    {
+      headers: {
+        'X-VTEX-API-AppKey': process.env.API_KEY ?? secrets.API_KEY,
+        'X-VTEX-API-AppToken': process.env.API_TOKEN ?? secrets.API_TOKEN,
+        'Content-Type': 'application/json',
+        Accept: 'application/json',
+      },
+    }
+  )
+
+  if (data) {
+    // there is any here because the result of this API is super big.
+    const ObjetosProntos = data.map((item: any) => {
+      const sellerD = item.items[0].sellers.filter(
+        (element: any) => element?.sellerDefault === true
+      )
+
+      const installments = sellerD?.map(
+        (el: any) => el?.commertialOffer?.Installments
+      )
+
+      const allInstallment: InstallmentProps[][] = []
+
+      const BasediscountHighlights = sellerD
+        ?.map((el: any) => el?.commertialOffer.DiscountHighLight)
+        .flat()
+
+      const discountHighlights = BasediscountHighlights.map((el: any) => {
+        return { name: Object.values(el).toString() }
+      })
+
+      installments?.forEach((element: any) => {
+        if (element !== undefined && element !== null) {
+          allInstallment.push(element)
+        }
+      })
+
+      const { pathname } = new URL(item?.link)
+
+      return {
+        productId: item.productId,
+        productName: item.productName,
+        link: pathname,
+        image: item.items[0].images[0]?.imageUrl,
+        listPrice: item.items[0].sellers[0]?.commertialOffer?.ListPrice,
+        price: item.items[0].sellers[0]?.commertialOffer?.Price,
+        brand: item.brand,
+        allInstallment: allInstallment.flat(),
+        discountHighlights,
+      }
+    })
+
+    return ObjetosProntos
+  }
+
+  return new GraphQLError('Não consultar os produtos da wishlist')
+}
+
 type NewsLetterVariable = {
   email: string
   id: string
@@ -353,8 +532,8 @@ async function newsLetter(_: unknown, { email }: NewsLetterVariable) {
     {
       headers: {
         'content-type': 'application/json',
-        'X-VTEX-API-APPTOKEN': process.env.API_TOKEN ?? '',
-        'X-VTEX-API-APPKEY': process.env.API_KEY ?? '',
+        'X-VTEX-API-APPTOKEN': process.env.API_TOKEN ?? secrets.API_TOKEN,
+        'X-VTEX-API-APPKEY': process.env.API_KEY ?? secrets.API_KEY,
       },
     }
   )