+++ CHANGELOG.md
@@ -13,6 +13,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - `SKU Selector` for work with multiple specifications ([#44](https://github.com/vtex-sites/accenturebb.store/pull/44)
 
 ### Fixed 
+- `Fix price` format ([#76](https://github.com/vtex-sites/accenturebb.store/pull/76)
 - `Fix layout` ([#71](https://github.com/vtex-sites/accenturebb.store/pull/71)
     - search filter remove product cluster id
     - blog section
@@ -28,6 +29,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - `Footer` new layout ([#32](https://github.com/vtex-sites/accenturebb.store/pull/32)
 
 ### Added
+- `Wishlist` ([#77](https://github.com/vtex-sites/accenturebb.store/pull/77))
 - `Newsletter` ([#75](https://github.com/vtex-sites/accenturebb.store/pull/75)
 - `Update repository` version 22.38.5 ([#65](https://github.com/vtex-sites/accenturebb.store/pull/65))
 - `Seller badge` on pdp ([#66](https://github.com/vtex-sites/accenturebb.store/pull/66)